refactor: refactor post routes into a separate module and integrate with main app
